training/requirements.txt
torchmetrics==1.4.1
onnx==1.16.2
onnxsim==0.4.36
onnx2tf==1.21.4
tensorflow==2.13.1
Pillow==10.4.0
